Ensure switch info is always available
For switches targetted via the `lampie.override` service, some
informaiton was missing in the internal state which would have caused
issues like failing to clear when a firmware duration expired on the
switch.

Also, fix an issue where errors would be logged for trying to handle
events on switches that are not tracked by the orchestrator.