Fix a rare issue when dismissing an override
If a switch was currently being used to display an override and a
notification with the exact same LED config, dismissing the override
would fail to issue commands to update the LEDs again to reflect that
the notification was still active.

This resolves that issue.

This change additionally adds some refactoring to make change processing
more consistent and easier to understand within the code base. The
underlying state of the switch is stored when the firmware updates the
LEDs. This allows the switch transition logic to always use the internal
state of the switch to determine if any updates actually need to be
made.